chore: code tidying and filterFn refactor
