+++ packages/@idemsInternational/gdrive-tools/package.json
@@ -5,7 +5,7 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "scripts": {
-    "start": "ts-node src/cli.ts",
+    "start": "ts-node src/commands/index.ts",
     "dev": "ts-node-dev --respawn --watch src/**/*.ts --",
     "build": "tsc -p ."
   },
