+++ packages/@idemsInternational/gdrive-tools/src/commands/index.ts
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 
 import { Command } from "commander";
-import downloadCmd from "./commands/download";
-import authorizeCmd from "./commands/authorize";
-import watchCmd from "./commands/watch";
-import { logProgramHelp } from "./utils";
+import downloadCmd from "./download";
+import authorizeCmd from "./authorize";
+import watchCmd from "./watch";
+import chalk from "chalk";
 
 const program = new Command();
 program.version("1.0.0").description("IDEMS Gdrive Tools");
@@ -20,4 +20,11 @@ if (require.main === module) {
     logProgramHelp(program);
   }
   program.parse(process.argv);
+}
+
+export function logProgramHelp(program: Command) {
+  console.log(chalk.yellow("No command specified. See help below:\n"));
+  program.outputHelp();
+  console.log("\n");
+  process.exit(0);
 }
