+++ packages/@idemsInternational/gdrive-tools/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./lib/authorize";
 export * from "./lib/download";
 export * from "./lib/watch";
+export * from "./models";
