+++ packages/@idemsInternational/gdrive-tools/src/models/index.ts
@@ -1,0 +1,13 @@
+/** File entry format from gdrive-downloader */
+export interface IGdriveEntry {
+  id: string;
+  name: string;
+  mimeType: string;
+  modifiedTime: string;
+  folderPath: string;
+  // only appears on xlsx uploaded files (not gsheet)
+  fullFileExtension?: string;
+  fileExtension?: string;
+  md5Checksum?: string;
+  size?: string;
+}
