+++ packages/@idemsInternational/gdrive-tools/src/utils/logging.utils.ts
@@ -1,6 +1,5 @@
 import boxen from "boxen";
 import chalk from "chalk";
-import { Command } from "commander";
 
 /** Record a 2-line error message in a box with additional optional logging and exit */
 export function logError(opts: Partial<ILogErrorOptions> = {}) {
@@ -14,13 +13,6 @@ export function logError(opts: Partial<ILogErrorOptions> = {}) {
   if (!logOnly) {
     process.exit(1);
   }
-}
-
-export function logProgramHelp(program: Command) {
-  console.log(chalk.yellow("No command specified. See help below:\n"));
-  program.outputHelp();
-  console.log("\n");
-  process.exit(0);
 }
 
 interface ILogErrorOptions {
